download non json content
when you 'GET' resource you know is not of type application/json or any text format the resource is handled by browser. For instance if u want to download resource with mime type application/octet-stream u want browser to ask user where to save. This commit make both that happen and handles error in case of failure.