Enable and fix existing test for serializeRes
The existing test for serializeRes did not return the Promise it
created, causing the test to pass in all cases due to not throwing an
exception.  Instead, return the Promise so the test will pass/fail based
on the promise resolution.

Instead of returning a Request object from the mocked fetch, let
fetch-mock build a Response with the desired headers.

Pass an object as the third parameter to fix `TypeError: Cannot read
property 'loadSpec' of undefined` due to the third argument being
undefined.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>