Fix tests for Node v4
For Node v4 npm installs separate copies of node-fetch for
isomorphic-fetch and fetch-mock.  This makes the `instanceof Headers`
check in the Headers constructor[1] fail for cross-module objects (such
as the `Headers` instance constructed in the test from isomorphic-fetch
which is passed to fetch-mock as a response.

Fix this by using plain objects to describe the response.

1.  https://github.com/bitinn/node-fetch/blob/v1.6.3/lib/headers.js#L22

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>