Use .buffer() for binary responses when .blob() is not available (#985)
* Enable and fix existing test for serializeRes

The existing test for serializeRes did not return the Promise it
created, causing the test to pass in all cases due to not throwing an
exception.  Instead, return the Promise so the test will pass/fail based
on the promise resolution.

Instead of returning a Request object from the mocked fetch, let
fetch-mock build a Response with the desired headers.

Pass an object as the third parameter to fix `TypeError: Cannot read
property 'loadSpec' of undefined` due to the third argument being
undefined.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

* Use .buffer as fallback when .blob unavailable

node-fetch version 1 does not implement support for .blob() on Response
objects.  In this case, use the non-standard .buffer() method which
returns a Node Buffer object.  Add tests to cover this case (and the
.text() analog for completeness).

Fixes #979

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

* Add default object for last serializeRes parameter

As requested.  This removes the need for passing an empty object
explicitly.

Remove it from the tests.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

* Fix tests for Node v4

For Node v4 npm installs separate copies of node-fetch for
isomorphic-fetch and fetch-mock.  This makes the `instanceof Headers`
check in the Headers constructor[1] fail for cross-module objects (such
as the `Headers` instance constructed in the test from isomorphic-fetch
which is passed to fetch-mock as a response.

Fix this by using plain objects to describe the response.

1.  https://github.com/bitinn/node-fetch/blob/v1.6.3/lib/headers.js#L22

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>