[PATCH v2] Include .response property on Errors as documented
When fetch fails as the result of a server error, callers may want to
examine the headers and/or content of the response to determine the
nature of the error.  Provide this information in the .response property
on the rejection Error object as currently documented in the HTTP Client
section of README.md.

Changes in v2:
- Apply responseInterceptor to error responses
- Add tests to cover responseInterceptor behavior on error responses

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>