fix(scheme): make scheme comparison case insensitive
As per RFC7235 auth scheme is case insensitive.

2.1. Challenge and Response

HTTP provides a simple challenge-response authentication framework
that can be used by a server to challenge a client request and by a
client to provide authentication information. It uses a case-
insensitive token as a means to identify the authentication scheme,
followed by additional information necessary for achieving.

https://tools.ietf.org/html/rfc7235#section-2.1

Refs #1531, #1473
Refs https://github.com/OAI/OpenAPI-Specification/issues/1876
Refs https://github.com/swagger-api/swagger-ui/issues/5965