chore(package): use more appropriate deps ranges
Note on deps starting with 0.x.y:

Hat matching may be considered broken because it
wont update ^0.1.2 to 0.2.0. When the software is emerging
use 0.x.y versions and hat matching will only match
the last varying digit (y). This is done on purpose.
The reason is that while the software is evolving
the API changes rapidly.