fix(resolver): bail out from all-off plugin prematurely
This bugfix solves nostalgic problem with reference resolution
manifesting as error: Cannot read property '0' of undefined.

There was an attempt to solve this in 98c9a1c54dee36e1f0a72905b0786b6e8cb59039
but some additional logic needed to be introduced to get rid of problem
for good.

Refs #1362