fix(resolver): allow removing Content-Type header from Response
Fetch API doesn't allow to mutate Response
or any associated objects like Headers.
Removing the Content-Type header allows us
to skip the JSON/YAML serialization.
To skip the serialization we have clone
the Response and Headers objects and remove
the Content-Type header in clones.

This doesn't apply to node-fetch@2 which
API allows to mutate Response and all
associated objects.

Refs https://github.com/swagger-api/swagger-ui/issues/8532